requirements.txt
ffmpeg
yt-dlp
brotli
python-dotenv